fix: pedometre izni + backend guvenlik/saglamlik/performans duzeltmeleri
Mobile:
- usePedometer: ACTIVITY_RECOGNITION runtime izni isteniyor (Android 10+ sensoru
  bunsuz susuyordu - adim sayar calismama bug'inin koku)
- getStepCountAsync Android'de desteklenmiyor; try/catch + canli watch'a gecis
- Veri kaybi korumasi: watch 0'dan baslayinca sunucudaki yuksek degeri ezmiyor

Backend:
- Rate limiting: auth genel (100/15dk) + login/register (20 basarisiz/15dk), trust proxy
- express-async-errors: async route hatalari artik errorHandler'a akiyor (asili istek yok)
- steps sync: GREATEST ile adim sayisi asla dusmuyor (race korumasi)
- challenge recalc: N+1 UPDATE -> tek sirali CTE UPDATE (performans)
- leaderboard rank: bigint -> int cast (frontend number bekliyor)
- kullanici aramada misafir hesaplar gizlendi

+++ backend/src/services/leaderboard.service.ts
@@ -24,7 +24,7 @@ export async function getDailyLeaderboard(userId: string): Promise<LeaderboardEn
        u.username,
        u.avatar_url AS "avatarUrl",
        COALESCE(ds.step_count, 0) AS "stepCount",
-       RANK() OVER (ORDER BY COALESCE(ds.step_count, 0) DESC) AS rank,
+       RANK() OVER (ORDER BY COALESCE(ds.step_count, 0) DESC)::int AS rank,
        u.id = $1 AS "isCurrentUser"
      FROM friends f
      JOIN users u ON u.id = f.friend_id
@@ -49,7 +49,7 @@ export async function getWeeklyLeaderboard(userId: string): Promise<LeaderboardE
        u.username,
        u.avatar_url AS "avatarUrl",
        COALESCE(SUM(ds.step_count), 0)::int AS "stepCount",
-       RANK() OVER (ORDER BY COALESCE(SUM(ds.step_count), 0) DESC) AS rank,
+       RANK() OVER (ORDER BY COALESCE(SUM(ds.step_count), 0) DESC)::int AS rank,
        u.id = $1 AS "isCurrentUser"
      FROM friends f
      JOIN users u ON u.id = f.friend_id
